webhook: prevent double node allocation (#75)
Add a validating webhook for the Cluster CR, which prevents an inventory
from being used in two different clusters.

Once a Cluster CR is created, all referenced inventories are assumed to
be occupied by that cluster. The creation of a second Cluster CR which
references any of the inventories, which the first Cluster CR already
occupies, is prevented.

related-to: harvester/harvester#8519

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>